+++ src/main/java/ch2/baseball/Baseball.java
@@ -1,4 +1,4 @@
-package ch2;
+package ch2.baseball;
 
 public class Baseball {
     private AnswerGenerator answerGenerator;
