+++ src/main/java/ch2/baseball/InputReceiver.java
@@ -1,4 +1,4 @@
-package ch2;
+package ch2.baseball;
 
 import java.util.Scanner;
