+++ src/test/java/ch2/InputValidatorTest.java
@@ -2,6 +2,7 @@ package ch2;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import ch2.baseball.InputValidator;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
